Tests the append blob function with a simple string (#431)
* Tests the append blob function with a simple string
- refactors/moves the model (concern) of the blobs module into models folder
* Removes expectest - so that we only depend on one testing framework
- uses assay to test instead

* Fixes the import to be local instead of public
- moves a function and its test nearest to the code that uses it
* fix whitespace in license header
